% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_accuracy.R
\name{f_accuracy}
\alias{f_accuracy}
\title{Accuracy}
\usage{
f_accuracy(x, y, c = 0.5)
}
\arguments{
\item{x}{vector with predictions}

\item{y}{vector with binary outcome}

\item{c}{cutoff. Should be between 0 and 1. Default is 0.5}
}
\description{
Accuracy
}
