% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load.R
\name{f_change}
\alias{f_change}
\title{This function is used to calculate price change in \%}
\usage{
f_change(x, from_start = FALSE, digits = 5)
}
\arguments{
\item{x}{stock price}

\item{from_start}{logical. If TRUE, change will be calculated since day 1}

\item{digits}{number of digits used when rounding the price change}
}
\value{
A vector with prince changes in \%
}
\description{
This function is used to calculate price change in \%
}
