% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_data_prep.R
\name{f_data_prep}
\alias{f_data_prep}
\title{Preparing data for training xgboost}
\usage{
f_data_prep(data)
}
\arguments{
\item{data}{data from f_load. Should be a data.table object}
}
\value{
data with target and indicators
}
\description{
Preparing data for training xgboost
}
