% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load_data.R
\name{f_load_data}
\alias{f_load_data}
\title{Load data}
\usage{
f_load_data(ticker, from_date = "2016-01-01", offline = FALSE)
}
\arguments{
\item{ticker}{ticker name}

\item{from_date}{load data since from date}

\item{offline}{logical. If TRUE, data will be loaded from data folder.}
}
\description{
Load data
}
