% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load.R
\name{f_load_fred}
\alias{f_load_fred}
\title{Loading data from Federal Reserve bank (FRED)}
\usage{
f_load_fred(symbol)
}
\arguments{
\item{symbol}{symbol to load}
}
\value{
A data.table
}
\description{
f_load_fred takes character value and loads data from Federal Reserve bank (FRED)
}
