% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load.R
\name{f_load_one}
\alias{f_load_one}
\title{This function loads one stock from Yahoo.}
\usage{
f_load_one(
  ticker,
  from_date = "2014-01-01",
  offline = FALSE,
  indicators = TRUE
)
}
\arguments{
\item{ticker}{ticker code from Yahoo. Use f_load to multiple stocks.}

\item{from_date}{loads data from the date untill today. Date format is yyyy-mm-dd.}

\item{offline}{logical. If TRUE, data will be loaded from data folder}

\item{indicators}{logical. If TRUE, technical indicators will be calculated as well. Default is TRUE}
}
\value{
A data.table
}
\description{
Get stock data from Yahoo by inputting ticker name.
}
