% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_trade.R
\name{f_plot_trade}
\alias{f_plot_trade}
\title{Plotting the trades}
\usage{
f_plot_trade(test_data, dt_transactions, mode = "candlesticks")
}
\arguments{
\item{test_data}{data.table with test data.}

\item{dt_transactions}{data.table with transactions. This should be outcome from f_trade.}

\item{mode}{plot mode. Should be "candlesticks" or "lines". Default is candlesticks}
}
\description{
Plotting the trades
}
