% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_train.R
\name{f_predict}
\alias{f_predict}
\title{Predicting from xgboost models}
\usage{
f_predict(model, data)
}
\arguments{
\item{model}{xgboost object obtained from f_train}

\item{data}{data to calculate predictions on. Should be a data.table object}
}
\value{
a numeric vector with predictions from the model
}
\description{
Predicting from xgboost models
}
