% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_report.R
\name{f_report}
\alias{f_report}
\title{Report from xgboost predictions}
\usage{
f_report(
  ticker,
  target_name = "target_day1",
  from_date = "2016-01-01",
  var_names = NULL,
  buy_cut = 0.6,
  sell_cut = 0.4,
  importance_plot = FALSE,
  shap_plot = FALSE,
  top_n = 7,
  offline = FALSE,
  mode = "candlesticks"
)
}
\arguments{
\item{ticker}{ticker name}

\item{target_name}{target name. See options in \code{f_target_names()}. Default is target_day1}

\item{var_names}{predictors used to train xgboost model. Default is NULL, meaning all predictors from f_indicators will be selected.}

\item{buy_cut}{cut off for when to buy based on predictions from xgboost. Should be between 0 and 1.}

\item{sell_cut}{cut off for when to sell based on predictions from xgboost. Should be between 0 and 1.}

\item{importance_plot}{logical. If TRUE, importance plot will be calculated. Default is FALSE.}

\item{shap_plot}{logical. If TRUE, shap plot will be calculated. Default is FALSE}

\item{top_n}{top n variables to show in importance plot.}

\item{offline}{logical. If TRUE, data will be loaded from data folder}

\item{mode}{plot mode. Should be "candlesticks" or "lines". Default is "candlesticks".}

\item{from}{date when loading data}
}
\description{
Report from xgboost predictions
}
