% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_split_train_test.R
\name{f_split_train_test}
\alias{f_split_train_test}
\title{Train test splitting data}
\usage{
f_split_train_test(data, split_pct = 0.8)
}
\arguments{
\item{data}{data}

\item{split_pct}{split percent. A number between 0 and 1. Default is a 80-20 split. An out of sample period is provided as well, so data is split into 80 training/test and 20 out of sample. The 80 training/tet is split into 80-20 training and 20 test}
}
\description{
Train test splitting data
}
