% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_step.R
\name{f_step}
\alias{f_step}
\title{Stepwise xgboost}
\usage{
f_step(data, target_name, init_var = NULL, var_names = NULL, n_max_steps = 10)
}
\arguments{
\item{data}{data.table or data.frame}

\item{target_name}{response variable name. See options in f_target_names()}

\item{init_var}{initial variables that always will be selected. Default is missing.}

\item{var_names}{feature names}

\item{n_max_steps}{maximum number of steps. Default is 10}
}
\description{
Stepwise xgboost
}
