% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_target_names.R
\name{f_target_names}
\alias{f_target_names}
\title{Target names}
\usage{
f_target_names()
}
\description{
\code{f_target_names} returns a character vector with possible target names.
The number indicates stock increase(= 1) and decrease (= 0) x days in future. So, target_day1 is a flag for whether or not the stock increased on day i + 1 looking at day i.
}
