% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_var_names.R
\name{f_var_names}
\alias{f_var_names}
\title{Variable names}
\usage{
f_var_names(data, var_names = NULL)
}
\arguments{
\item{data}{data input is used to only return variables present in data}

\item{var_names}{variable names. Default is NULL, which means all predictor names from f_indicators that are present in data are returned. If a vector of variable names are supplied, f_var_names will return the vector}
}
\description{
Variable names
}
