% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_weight_observations.R
\name{f_weight_observations}
\alias{f_weight_observations}
\title{Weight observations with exponential decay}
\usage{
f_weight_observations(n)
}
\arguments{
\item{n}{number of rows in data}
}
\description{
Weight observations with exponential decay
}
