% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{plot_shap}
\alias{plot_shap}
\title{plot shap}
\usage{
plot_shap(xgb_model, mat_train, top_n = 7)
}
\arguments{
\item{xgb_model}{xgboost object}

\item{mat_train}{train data as a matrix}

\item{top_n}{top n most influential variables to print}
}
\description{
plot shap
}
