% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{shap_prep}
\alias{shap_prep}
\title{prep shap data}
\usage{
shap_prep(shap = shap_result, X_train = mat_train, top_n)
}
\arguments{
\item{shap}{shap_result}

\item{X_train}{train data as a matrix}

\item{top_n}{top n variables to print in plot}
}
\description{
prep shap data
}
