% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{shap_score_rank}
\alias{shap_score_rank}
\title{return matrix of shap score and mean ranked score list}
\usage{
shap_score_rank(xgb_model = xgb_mod, shap_approx = TRUE, X_train = mat_train)
}
\arguments{
\item{xgb_model}{xgboost object}

\item{shap_approx}{logical}

\item{X_train}{train data as a matrix}
}
\description{
return matrix of shap score and mean ranked score list
}
