% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{std1}
\alias{std1}
\title{a function to standardize feature values into same range}
\usage{
std1(x)
}
\arguments{
\item{x}{vector}
}
\description{
a function to standardize feature values into same range
}
