% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{var_importance}
\alias{var_importance}
\title{Importance plot}
\usage{
var_importance(shap_result, top_n = 10)
}
\arguments{
\item{shap_result}{shap_result}

\item{top_n}{top n variables to print in plot}
}
\description{
Importance plot
}
